feat: movies page implemented (will be improved later)

+++ back/importer/src/main/java/cinema/importer/models/Movie.java
@@ -20,5 +20,5 @@ public class Movie {
     String imageUrl;
     String crew;
     double imDbRating;
-    double imDbRatingCount;
+    int imDbRatingCount;
 }
